requirements.txt
# Core dependencies
tree-sitter>=0.20.0
tree-sitter-languages>=1.7.0

# Tree-sitter language parsers
tree-sitter-c-sharp>=0.20.0
tree-sitter-bash>=0.20.0
tree-sitter-haskell>=0.14.0
tree-sitter-scala>=0.19.0
tree-sitter-lua>=0.0.14
tree-sitter-yaml>=0.5.0
tree-sitter-toml>=0.5.1
tree-sitter-json>=0.19.0
tree-sitter-markdown>=0.3.2
tree-sitter-csv>=0.1.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.1
pytest-benchmark>=4.0.0
pytest-xdist>=3.3.1
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
coverage[toml]>=7.3.0
hypothesis>=6.82.0

# Cache dependencies (optional)
redis>=4.5.0

# Semantic search dependencies (dormant features)
voyageai>=0.2.0
qdrant-client>=1.7.0

# Security dependencies
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.4

# Testing utilities
factory-boy>=3.3.0
faker>=19.6.1
responses>=0.23.3
freezegun>=1.2.2

# Development tools
ipython>=8.14.0
jupyter>=1.0.0
ipdb>=0.13.13
rich>=13.5.2

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Security scanning
safety>=2.3.5
bandit>=1.7.5

# Performance profiling
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.1
numpy>=1.24.0
jinja2>=3.1.2

# Metrics and monitoring
psutil>=5.9.0
prometheus-client>=0.17.0

# API testing
httpx>=0.24.1
pytest-httpx>=0.22.0

# Async testing support
aiofiles>=23.2.1
asynctest>=0.13.0

# WebSocket support
aiohttp>=3.9.0
websockets>=12.0

# Code quality
pylint>=2.17.5
pydocstyle>=6.3.0
radon>=6.0.1

# Type stubs
types-requests>=2.31.0
types-setuptools>=68.1.0

jsonschema>=4.17.0

# Repository management dependencies (standard library - listed for completeness)
# sqlite3 - Database operations (standard library)
# json - JSON serialization (standard library) 
# datetime - Date/time handling (standard library)
# pathlib - Path operations (standard library)

# Additional MCP dependencies
pydantic>=2.0.0
click>=8.0.0
watchdog>=4.0.0
